+++ src/client/utils/puzzle-state-management-functions/allSquareIdsAndPuzzleVals.ts
