+++ src/client/utils/puzzle-state-management-functions/initialSquareStatePopulation.ts
