+++ src/client/utils/puzzle-state-management-functions/isPuzzleFinished.ts
