+++ src/client/utils/puzzle-state-management-functions/newFilledSquare.ts
