+++ src/client/utils/puzzle-state-management-functions/puzzleStringValidation.ts
