+++ src/client/utils/puzzle-state-management-functions/puzzleStringsFromSquares.ts
