+++ src/client/utils/puzzle-state-management-functions/puzzleValueChange.ts
