+++ src/client/utils/puzzle-state-management-functions/squaresFromPuzzleStrings.ts
