+++ src/client/utils/puzzle-state-management-functions/updateSquaresDuplicates.ts
